requirements.txt
flask==3.0.0
openai==1.3.0
python-dateutil==2.8.2
hypothesis==6.92.0
pytest==7.4.3
requests==2.31.0
python-dotenv==1.0.0
